backend/requirements.txt
flask==3.1.0
flask-pymongo==2.3.0
flask-jwt-extended==4.7.1
flask-cors==5.0.1
kerykeion==5.7.2
openai==1.61.0
python-dotenv==1.0.1
gunicorn==23.0.0
pymongo[srv]==4.11.3
bcrypt==4.2.1
